validasi form suksess

+++ app/Controllers/KosanController.php
@@ -21,15 +21,17 @@ class KosanController extends BaseController
     // INDEX
     public function index()
     {
-        
+
         return view('auth/admin/data_kosan_page');
     }
 
     // CREATE DATA KOSAN
     public function create()
     {
+        session();
         $data = [
             'title' => 'Form Tambah Data Kosan | Owner',
+            'validation' => \Config\Services::validation(),
         ];
 
         return view('auth/owner/tambah_kosan_page', $data);
@@ -38,7 +40,6 @@ class KosanController extends BaseController
     // SAVE CREATE DATA KOSAN
     public function save()
     {
-        // Ambil data kosan dari input
         $data_kosan = [
             'namaKost' => $this->request->getVar('namaKost'),
             'alamat' => $this->request->getVar('alamat'),
@@ -53,64 +54,75 @@ class KosanController extends BaseController
         ];
 
         // validasi data kosan
-        // $validated = [
-        //     'namaKost' => [
-        //         'rules' => 'required|min_length[5]|max_length[50]',
-        //         'errors' => [
-        //             'required' => 'Nama kosan harus diisi',
-        //             'min_length' => 'Nama kosan minimal 5 karakter',
-        //             'max_length' => 'Nama kosan maksimal 50 karakter',
-        //         ]
-        //     ],
-        //     'alamat' => [
-        //         'rules' => 'required|min_length[5]',
-        //         'errors' => [
-        //             'required' => 'Alamat kosan harus diisi',
-        //             'min_length' => 'Alamat kosan minimal 5 karakter',
-        //         ]
-        //     ],
-        //     'kota' => [
-        //         'rules' => 'required|inlist[Lampung Barat,Tanggamus,Lampung Selatan,Lampung Timur,Lampung Tengah,Lampung Utara,Way Kanan,Pesawaran,Tulang Bawang Barat,Tulang Bawang,Pesisir Barat,Bandar Lampung,Metro]',
-        //         'errors' => [
-        //             'required' => 'Kecamatan kosan harus diisi',
-        //             'inlist' => 'Kota/Kabupaten kosan harus dipilih dengan benar',
-        //         ]
-        //     ],
-        //     'deskripsi' => [
-        //         'rules' => 'required|min_length[5]',
-        //         'errors' => [
-        //             'required' => 'Deskripsi kosan harus diisi',
-        //             'min_length' => 'Deskripsi kosan minimal 5 karakter',
-        //         ]
-        //     ],
-        //     'fasilitas' => [
-        //         'rules' => 'required',
-        //         'errors' => [
-        //             'required' => 'Fasilitas kosan harus diisi',
-        //         ]
-        //     ],
-        //     'harga' => [
-        //         'rules' => 'required|numeric',
-        //         'errors' => [
-        //             'required' => 'Harga kosan harus diisi',
-        //             'numeric' => 'Harga kosan harus berupa angka',
-        //         ]
-        //     ],
-        //     'type' => [
-        //         'rules' => 'required|inlist[Putra,Putri]',
-        //         'errors' => [
-        //             'required' => 'Type kosan harus diisi',
-        //             'inlist' => 'Type kosan harus dipilih dengan benar',
-        //         ]
-        //     ],
-        // ];
+        $validated = $this->validate(
+            [
+                'namaKost' => [
+                    'rules' => 'required|min_length[5]|max_length[50]',
+                    'errors' => [
+                        'required' => 'Nama kosan harus diisi',
+                        'min_length' => 'Nama kosan minimal 5 karakter',
+                        'max_length' => 'Nama kosan maksimal 50 karakter',
+                    ]
+                ],
+                'alamat' => [
+                    'rules' => 'required|min_length[5]',
+                    'errors' => [
+                        'required' => 'Alamat kosan harus diisi',
+                        'min_length' => 'Alamat kosan minimal 5 karakter',
+                    ]
+                ],
+                'kota' => [
+                    'rules' => 'required|in_list[Lampung Barat,Tanggamus,Lampung Selatan,Lampung Timur,Lampung Tengah,Lampung Utara,Way Kanan,Pesawaran,Tulang Bawang Barat,Tulang Bawang,Pesisir Barat,Bandar Lampung,Metro]',
+                    'errors' => [
+                        'required' => 'Kecamatan kosan harus diisi',
+                        'in_list' => 'Kota/Kabupaten kosan harus dipilih dengan benar',
+                    ]
+                ],
+                'deskripsi' => [
+                    'rules' => 'required|min_length[5]',
+                    'errors' => [
+                        'required' => 'Deskripsi kosan harus diisi',
+                        'min_length' => 'Deskripsi kosan minimal 5 karakter',
+                    ]
+                ],
+                'fasilitas' => [
+                    'rules' => 'required',
+                    'errors' => [
+                        'required' => 'Fasilitas kosan harus diisi',
+                    ]
+                ],
+                'harga' => [
+                    'rules' => 'required|numeric',
+                    'errors' => [
+                        'required' => 'Harga kosan harus diisi',
+                        'numeric' => 'Harga kosan harus berupa angka',
+                    ]
+                ],
+                'type' => [
+                    'rules' => 'required|in_list[Putra,Putri]',
+                    'errors' => [
+                        'required' => 'Type kosan harus diisi',
+                        'in_list' => 'Type kosan harus dipilih dengan benar',
+                    ]
+                ],
+                'foto_1' => [
+                    'rules' => 'uploaded[foto_1]|max_size[foto_1,1024]|is_image[foto_1]|mime_in[foto_1,image/jpg,image/jpeg,image/png]',
+                    'errors' => [
+                        'uploaded' => 'Foto kosan harus diisi',
+                        'max_size' => 'Ukuran foto kosan maksimal 1 MB',
+                        'is_image' => 'File yang diupload harus berupa gambar',
+                        'mime_in' => 'File yang diupload harus berupa gambar',
+                    ]
+                ],
+            ]
+        );
 
         // // jika validasi error
-        // if (!$validated) {
-        //     return redirect()->to('/onlyOwners/create')->withInput();
-        // }
+        if ($validated == FALSE) {
+            return redirect()->back()->withInput();
+        }
 
-        // Insert data kosan ke database
+        //Insert data kosan ke database
         $this->kosanModel->insert($data_kosan);
 
         // mengambil id kosan yang baru saja diinputkan
@@ -141,7 +153,7 @@ class KosanController extends BaseController
 
             $this->fotoKosanModel->insert($data_foto_kosan);
 
-            $foto_2->move( '../public/foto_kosan', $name_foto2);
+            $foto_2->move('../public/foto_kosan', $name_foto2);
         }
 
         if ($foto_3->getName() !== '') {
@@ -199,18 +211,86 @@ class KosanController extends BaseController
             'idPemilik' => user_id(),
             'updated_at' => date('Y-m-d H:i:s'),
         ];
+        $validated = $this->validate(
+            [
+                'namaKost' => [
+                    'rules' => 'required|min_length[5]|max_length[50]',
+                    'errors' => [
+                        'required' => 'Nama kosan harus diisi',
+                        'min_length' => 'Nama kosan minimal 5 karakter',
+                        'max_length' => 'Nama kosan maksimal 50 karakter',
+                    ]
+                ],
+                'alamat' => [
+                    'rules' => 'required|min_length[5]',
+                    'errors' => [
+                        'required' => 'Alamat kosan harus diisi',
+                        'min_length' => 'Alamat kosan minimal 5 karakter',
+                    ]
+                ],
+                'kota' => [
+                    'rules' => 'required|in_list[Lampung Barat,Tanggamus,Lampung Selatan,Lampung Timur,Lampung Tengah,Lampung Utara,Way Kanan,Pesawaran,Tulang Bawang Barat,Tulang Bawang,Pesisir Barat,Bandar Lampung,Metro]',
+                    'errors' => [
+                        'required' => 'Kecamatan kosan harus diisi',
+                        'in_list' => 'Kota/Kabupaten kosan harus dipilih dengan benar',
+                    ]
+                ],
+                'deskripsi' => [
+                    'rules' => 'required|min_length[5]',
+                    'errors' => [
+                        'required' => 'Deskripsi kosan harus diisi',
+                        'min_length' => 'Deskripsi kosan minimal 5 karakter',
+                    ]
+                ],
+                'fasilitas' => [
+                    'rules' => 'required',
+                    'errors' => [
+                        'required' => 'Fasilitas kosan harus diisi',
+                    ]
+                ],
+                'harga' => [
+                    'rules' => 'required|numeric',
+                    'errors' => [
+                        'required' => 'Harga kosan harus diisi',
+                        'numeric' => 'Harga kosan harus berupa angka',
+                    ]
+                ],
+                'type' => [
+                    'rules' => 'required|in_list[Putra,Putri]',
+                    'errors' => [
+                        'required' => 'Type kosan harus diisi',
+                        'in_list' => 'Type kosan harus dipilih dengan benar',
+                    ]
+                ],
+                'foto_1' => [
+                    'rules' => 'uploaded[foto_1]|max_size[foto_1,1024]|is_image[foto_1]|mime_in[foto_1,image/jpg,image/jpeg,image/png]',
+                    'errors' => [
+                        'uploaded' => 'Foto kosan harus diisi',
+                        'max_size' => 'Ukuran foto kosan maksimal 1 MB',
+                        'is_image' => 'File yang diupload harus berupa gambar',
+                        'mime_in' => 'File yang diupload harus berupa gambar',
+                    ]
+                ],
+            ]
+        );
+
+        // // jika validasi error
+        if ($validated == FALSE) {
+            return redirect()->back()->withInput();
+        }
+        
         $this->kosanModel->update($this->request->getVar('id_kosan'), $data);
         if ($this->request->getFile('foto_1')->getName() !== '') {
 
             $foto = $this->request->getFile('foto_1');
             $namaFoto = $foto->getRandomName();
-            unlink( '../public/foto_kosan/' . $this->fotoKosanModel->where('id_kosan', $this->request->getVar('id_kosan'))->findAll()[0]['nama_foto']);
+            unlink('../public/foto_kosan/' . $this->fotoKosanModel->where('id_kosan', $this->request->getVar('id_kosan'))->findAll()[0]['nama_foto']);
             $data_foto = [
                 'nama_foto' => $namaFoto,
                 'id_kosan' => $this->request->getVar('id_kosan'),
             ];
             $this->fotoKosanModel->update($this->fotoKosanModel->where('id_kosan', $this->request->getVar('id_kosan'))->findAll()[0]['id_foto'], $data_foto);
-            $foto->move( '../public/foto_kosan', $namaFoto);
+            $foto->move('../public/foto_kosan', $namaFoto);
         }
         if ($this->request->getFile('foto_2')->getName() !== '') {
             $foto = $this->request->getFile('foto_2');
@@ -220,12 +300,12 @@ class KosanController extends BaseController
                 'id_kosan' => $this->request->getVar('id_kosan'),
             ];
             if ($jumlahFoto > 1) {
-                unlink( '../public/foto_kosan/' . $this->fotoKosanModel->where('id_kosan', $this->request->getVar('id_kosan'))->findAll()[1]['nama_foto']);
+                unlink('../public/foto_kosan/' . $this->fotoKosanModel->where('id_kosan', $this->request->getVar('id_kosan'))->findAll()[1]['nama_foto']);
                 $this->fotoKosanModel->update($this->fotoKosanModel->where('id_kosan', $this->request->getVar('id_kosan'))->findAll()[1]['id_foto'], $data_foto);
             } else {
                 $this->fotoKosanModel->insert($data_foto);
             }
-            $foto->move( '../public/foto_kosan', $namaFoto);
+            $foto->move('../public/foto_kosan', $namaFoto);
         }
         if ($this->request->getFile('foto_3')->getName() !== '') {
             $foto = $this->request->getFile('foto_3');
@@ -235,13 +315,13 @@ class KosanController extends BaseController
                 'id_kosan' => $this->request->getVar('id_kosan'),
             ];
             if ($jumlahFoto > 2) {
-                unlink( '../public/foto_kosan/' . $this->fotoKosanModel->where('id_kosan', $this->request->getVar('id_kosan'))->findAll()[2]['nama_foto']);
+                unlink('../public/foto_kosan/' . $this->fotoKosanModel->where('id_kosan', $this->request->getVar('id_kosan'))->findAll()[2]['nama_foto']);
                 $this->fotoKosanModel->update($this->fotoKosanModel->where('id_kosan', $this->request->getVar('id_kosan'))->findAll()[2]['id_foto'], $data_foto);
             } else {
                 $this->fotoKosanModel->insert($data_foto);
             }
 
-            $foto->move( '../public/foto_kosan', $namaFoto);
+            $foto->move('../public/foto_kosan', $namaFoto);
         }
         return redirect()->to('/owner/kosan_anda');
     }
